+++ domain-resolv/src/lookup/addr.rs
@@ -8,8 +8,7 @@ use domain_core::bits::message::RecordIter;
 use domain_core::bits::name::{Dname, DnameBuilder, ParsedDname};
 use domain_core::iana::Rtype;
 use domain_core::rdata::parsed::Ptr;
-use ::conf::ResolvOptions;
-use ::resolver::{Answer, Query, Resolver};
+use crate::resolver::Resolver;
 
 
 //------------ lookup_addr ---------------------------------------------------
@@ -22,8 +21,8 @@ use ::resolver::{Answer, Query, Resolver};
 /// 
 /// The value returned upon success can be turned into an iterator over
 /// host names via its `iter()` method. This is due to lifetime issues.
-pub fn lookup_addr(resolv: &Resolver, addr: IpAddr) -> LookupAddr {
-    let name = dname_from_addr(addr, resolv.options());
+pub fn lookup_addr<R: Resolver>(resolv: &R, addr: IpAddr) -> LookupAddr<R> {
+    let name = dname_from_addr(addr);
     LookupAddr(resolv.query((name, Rtype::Ptr)))
 }
 
@@ -33,10 +32,10 @@ pub fn lookup_addr(resolv: &Resolver, addr: IpAddr) -> LookupAddr {
 /// The future for [`lookup_addr()`].
 ///
 /// [`lookup_addr()`]: fn.lookup_addr.html
-pub struct LookupAddr(Query);
+pub struct LookupAddr<R: Resolver>(R::Query);
 
-impl Future for LookupAddr {
-    type Item = FoundAddrs;
+impl<R: Resolver> Future for LookupAddr<R> {
+    type Item = FoundAddrs<R>;
     type Error = io::Error;
 
     fn poll(&mut self) -> Poll<Self::Item, Self::Error> {
@@ -51,19 +50,21 @@ impl Future for LookupAddr {
 ///
 /// The only purpose of this type is to return an iterator over host names
 /// via its `iter()` method.
-pub struct FoundAddrs(Answer);
+pub struct FoundAddrs<R: Resolver>(R::Answer);
 
-impl FoundAddrs {
+impl<R: Resolver> FoundAddrs<R> {
     /// Returns an iterator over the host names.
     pub fn iter(&self) -> FoundAddrsIter {
         FoundAddrsIter {
-            name: self.0.canonical_name(),
-            answer: self.0.answer().ok().map(|sec| sec.limit_to::<Ptr>())
+            name: self.0.as_ref().canonical_name(),
+            answer: {
+                self.0.as_ref().answer().ok().map(|sec| sec.limit_to::<Ptr>())
+            }
         }
     }
 }
 
-impl IntoIterator for FoundAddrs {
+impl<R: Resolver> IntoIterator for FoundAddrs<R> {
     type Item = ParsedDname;
     type IntoIter = FoundAddrsIter;
 
@@ -72,7 +73,7 @@ impl IntoIterator for FoundAddrs {
     }
 }
 
-impl<'a> IntoIterator for &'a FoundAddrs {
+impl<'a, R: Resolver> IntoIterator for &'a FoundAddrs<R> {
     type Item = ParsedDname;
     type IntoIter = FoundAddrsIter;
 
@@ -113,10 +114,10 @@ impl Iterator for FoundAddrsIter {
 //------------ Helper Functions ---------------------------------------------
 
 /// Translates an IP address into a domain name.
-fn dname_from_addr(addr: IpAddr, opts: &ResolvOptions) -> Dname {
+fn dname_from_addr(addr: IpAddr) -> Dname {
     match addr {
         IpAddr::V4(addr) => dname_from_v4(addr),
-        IpAddr::V6(addr) => dname_from_v6(addr, opts)
+        IpAddr::V6(addr) => dname_from_v6(addr)
     }
 }
 
@@ -135,23 +136,22 @@ fn dname_from_v4(addr: Ipv4Addr) -> Dname {
 ///
 /// As there are several ways to do this, the functions depends on
 /// resolver options, namely `use_bstring` and `use_ip6dotin`.
-fn dname_from_v6(addr: Ipv6Addr, opts: &ResolvOptions) -> Dname {
+fn dname_from_v6(addr: Ipv6Addr) -> Dname {
     let mut res = DnameBuilder::new();
-    for item in addr.segments().iter().rev() {
-        let text = format!("{:04x}", item);
-        let text = text.as_bytes();
-        res.append_label(&text[3..4]).unwrap();
-        res.append_label(&text[2..3]).unwrap();
-        res.append_label(&text[1..2]).unwrap();
-        res.append_label(&text[0..1]).unwrap();
+    for &item in addr.octets().iter().rev() {
+        res.append_label(&[hexdigit(item >> 4)]).unwrap();
+        res.append_label(&[hexdigit(item)]).unwrap();
     }
     res.append_label(b"ip6").unwrap();
-    if opts.use_ip6dotint {
-        res.append_label(b"int").unwrap();
-    }
-    else {
-        res.append_label(b"arpa").unwrap();
-    }
+    res.append_label(b"arpa").unwrap();
     res.into_dname().unwrap()
+}
+
+fn hexdigit(nibble: u8) -> u8 {
+    match nibble % 0x0F {
+        0...10 => nibble + b'0',
+        10...16 => nibble - 10 + b'a',
+        _ => unreachable!()
+    }
 }